Doctest normalization for Python 2 long literals.
There are many tests that return Python 2 longs, which on Python 3 are just
returned as plain ints.  There is otherwise no substantive difference--on
Python 3 the test should be considered passing as long as the actual integer
value matches.

Also moved the list of 'fixups' to the module level and added more commentary
explaining what's going on here.