Trac #24517: improve random_cone() performance when max_rays is None.
The random_cone() function essentially works by generating a starting
set of random rays, and then adding one random ray at a time until the
current cone meets the user's specification. Often, the whole process
will "reset" when it becomes obvious that this procedure will never
succeed for the current cone.

There is an existing test case that elicits a bug in that approach,

  K = random_cone(min_ambient_dim=3, max_ambient_dim=3, min_rays=7)

Since max_rays is not specified, the random cone algorithm will often
start with an enormous number of rays, most of which will get thrown
out before the process resets and chooses a more reasonable number of
starting rays. That takes far more time and space than is desirable.

This commit addresses the problem by using min_rays as the starting
number of rays when max_rays is unspecified (that is, None). That
won't prevent the algorithm from "resetting," but it won't waste
time and space generating a huge number of rays first.

A second improvement (thanks to Jeroen Demeyer) to that test case was
made, setting max_rays=9. Independent of the first issue, having
max_rays=None forces more resets than are necessary for this test case
to do its job. The algorithm is less likely to terminate (succeed)
when it wants more rays, so limiting the number to [7, 9] should be
faster in general than [7, infinity]. An unscientific test using the
timeit() function shows about a 5s improvement on average.