Trac #24517: simplify the recent random_cone() performance fix.
A recent commit improves the performance of random_cone() when max_rays
is None by adding a special case: when max_rays is None (that is,
unbounded), we choose a smaller number of starting rays. The
conditional is used because there is another performance issue to keep
in mind: we don't want to start with too few rays, and have to add
them one-at-a-time in a larger space, where the rays won't be
redundant.

The point of diminishing returns is when the number of rays is around
twice the ambient dimension, because, for example, six rays can
generate all of three-dimensional space, making any additional rays
redundant. Since it is extremely unlikely that more than 2*d rays will
be useful, we can unconditionally cap the number of starting rays at
2*d, rather than worry about whether or not max_rays was specified.

This should still fix the issue reported in Trac #24517, but performs
better when the user specifies a large number for max_rays in a small
ambient space. For example, with the original fix in 6f9be50,

  K = random_cone(min_ambient_dim=1,max_ambient_dim=1,max_rays=10000000)

could crash if we tried to construct r=10000000 initial rays in some
iteration. Now that number would be capped to min(2*d,r) = 2.

This has a side effect, namely that if you ask for more than 2*d rays,
then you'll probably get the minimum amount, because we'll start with
2*d and add them one-at-a-time.