Fix several tests in this module.
The one test that tested the actual bytes of a pickle fails due to minor
differences in how the pickle format is written, that are inconsequential
to deserialization but still result in slightly different byte contents.
So instead I replaced that with a 'disassembly', as the actual opcodes are
still the same.