Stricter locking around sage-rebase on Cygwin
Now we don't allow sage-rebase to run while *any* package is still being
installed; as this appears to cause problems sometimes.

Unfortunately, this has a pretty negative impact on the efficiency of
parallel builds on Cygwin, but I think it's a necessary sacrifice.  I think
later it might be possible to reduce this impact somewhat (e.g. it's not
strictly necessary to run sage-rebase for *every* package)