stop considering inexact balls equal to themselves
Sage's implementation of comparison in real and complex ball fields used
to make all elements (even non-point intervals) equal to themselves,
because the documentation of arb specifies that “Identical pointers are
understood to give permission for algebraic simplification”. In spite of
this provisio, arb itself now provides comparison functions with x != x
for inexact x. This commit makes the corresponding change in Sage for
consistency. I wouldn't consider this a break of compatibility since, in
geneal, x != y returning False for two balls x, y does not guarantee
that the corresponding reals are distinct.