py3: Fixes for sage.repl.rich_output
Mostly str/bytes fixes, as can be imagined.  Added an OutputBuffer.get_str
method which on Python 2 return OutputBuffer.get(), and on Python 3 returns
OutputBuffer.get_unicode().

Changed several of the tests to use .get_str() instead of .get().  I believe
this is justifiable, in that on Python 2 it is just directly calling .get(),
so the test is still the same for Python 2, but also has consistent results
on Python 3.