py3: minor test fixes in sage.libs.mpmath.ext_impl
These failures are due to differences between Python 2/3 in how many digits are
displayed by default when printing floats, as opposed to just displaying their
repr.  In these cases there's no particular reason to *print* instead of just
show the repr, so we do the latter which is consistent between Pythons.