Provide a workaround for https://trac.sagemath.org/ticket/24508
This simply copy/pastes the random.Random implementation from Python 2, with
minor tweaks for Python 3 syntax compatibility.  We use this only for
running the tests in Python 3 so that code that uses random.Random gives
consistent results in the test suite.

Once Sage switches to Python 3 by default this should be removed, and the
affected tests systematically updated.