py3: fix sage.interfaces.sage0 to better handle file closing (particularly to avoid ResourceWarnings)
The pickle loading code is now wrapped in functions that are added to the
init_code, which handle opening and reading pickles within a context manager.
This also cleans up the init_code argument handling a bit to ensure that code
required for the class to work at all is always added after any user-specified
init_code.