Changed a bit how SAGE_SPKG_CONFIGURE works:
Now, instead of being a macro that directly subsitutes the individual spkg checks,
it's actually a higher-order macro that defines macros named SAGE_SPKG_CONFIGURE_<PKGNAME>.

All of these macros are included in m4/sage_spkg_configures.ac so that they all get expanded,
but the advantage of this is that individual packages' SAGE_SPKG_CONFIGURE can use
AC_REQUIRE to ensure that other packages are configured as a prerequisite.

This in gfortran to ensure that gcc is always configured first.  Then we can disable
gfortran if gcc is going to be installed.