Minor fixes to sage.tests.cmdline to work on Python 3.
In this case it takes advantage of the encoding keyword argument to Popen
available on Python 3 to handle encoding/decoding bytes (this also replaces
direct use of os.read but this only has minor overhead).

Not all the tests in this module pass yet, unsurprisingly, as they test
many parts of Sage.  But at least the process runner itself will work now.