Updates several miscellaneous packages for #24024:
* gfan
* lcalc
* ratpoints
* rubiks
* symmetrica

These packages don't specifically have anything in common here except for
having somewhat unusual custom install scripts, but that aren't particularly
complicated either.

In the case of rubiks we removed the old code for removing 'bin/cube': this
is long-since obsolete and would not be the appropriate way to handle a
file conflict between packages.