Fixing that 1/CIF(RIF(-1,1),0) would result in an infinite recursion in ComplexIntervalFieldElement.__invert__().
Fixing that this method leaked memory (through its early return
statements not calling mpfr_clear) and improving performance.

This is done by doing the necessary transformations to bring the
complex interval into standard form in place. This avoids the pattern
of the method calling itself all together in favor of being more
explicit.

The transformations to bring it into standard form (where the
interval for real part contains a non-negative number and the
imaginary part is non-negative) are different, but easier than
the one in Rokne-Lancaster.
By doing them in place (using mpfr_swap where possible), we avoid
constructing a python ComplexIntervalFieldElement to do the
multiplication by powers of I.