Upgrade cddlib
and use cddexec instead of cdd_both_reps. cddexec should be faster and is,
according to upstream, "done the right way". The change to cddexec makes some
numerical issues visible, so some things that worked before are not supported
anymore. Essentially, we now require an inexact redundancy-free representation
to be convertible back and forth between V and H representation (without loss
of vertices/inequalities) before we trust it. Some standard examples, fail this
check. I can not really make the call on whether these are bugs in cddlib or
numerical issues that could not possibly be fixed (as I do not understand this
well enough.)

So why this change? Mostly because we do not patch cddlib anymore and just use
it as is. It is also faster in some cases (that's what upstream cared most
about when the cddlib author heard about our cdd_both_reps,) e.g.,

%timeit polytopes.six_hundred_cell(exact=False)
went from 993ms to 246ms

Some other examples see minor slowdowns because of the additional
transformation that we compute initially to make sure that the conversion back
and forth worked out.