py3: adds a _import_module_from_path helper function
this implements the same routine that was used here to import the Python
module compiled from the FORTRAN code, but it has both Python 2 and Python3
implementations, both of which are very different (the Python 2 version does
work on Python 3, but uses deprecated interfaces).

In principle this function could be useful elsewhere, but this is the only
place I found that uses it currently, so I kept it in this module.