Skip publication steps if SECRET_DOCKER_PASS has not been set
Note that there is no way to check whether both SECRET_DOCKER_PASS &&
DOCKER_USER have been set. Also, this is an alpha feature of GitLab, so the
syntax might change. As an alternative we could check this at the beginning of
the script block and just "exit 0"; but currently a lot of time and in
particular bandwidth is wasted by pulling the docker images and then not
pushing them because of missing credentials.