uninstall should happen before build
This is important for two reasons:

a) When upgrading from a 'legacy' install, the spkg-legacy-uninstall scripts
   often need a chance to run before building the new version of a package
   (e.g. if having an old version of the package installed interferes with
   building the new version, as with Python)

b) When downgrading from a new-style install to an old-style install (without
   a file manifest) running uninstall after install can break the
   just-installed package by removing some of its files from SAGE_LOCAL

This change has the downside that if building a new version of a package
fails, one is left with a broken build (as opposed to waiting to uninstall
an old version until the new version builds successfully).  But that's at
least no different from the status quo, and will be rectified once
https://trac.sagemath.org/ticket/25202 is implemented (the ability to
backup existing installed packages and restore them upon upgrade failure).