py3: map/filter fixes for sage.graphs.generators.classical_geometries
Heavy use of map/filter in this module leads to lots of errors on Python 3
since they are then passed to interfaces that expect lists not unsized
arbitrary iterables.

Many of those interfaces should probably be improved, but in the meantime
we can get a lot further by just avoiding map/filter where appropriate.