Allow repeated key polynomials
the comment in the code was only true for a full run of mac_lane_approximants
starting from a GaussValuation. That's the code I had this originally written
for. I do not think that the block is necessary (it might have been in some
broken p-adics setting but if it is then, that should be fixed in the p-adics
code.)