SageMath is also available on NixOS and conda
also some wording changes where I stumbled upon sentences. I am not sure what
to do here with the Sage vs SageMath thing. I had thought that Sage is the name
of the Python library and SageMath is the distribution and the project. So
shouldn't it be SageMath almost everywhere?