py3: sort the output from symmetrized_coordinate_sums so that it is deterministic on Python 2 and 3
py3: even this test class needs a __hash__ in order for the tests to work right

py3: some items/iteritems fixes

for the items() test better to test the behavior (the returned value is an iterator
and hence supports next()) rather than the exact type repr which is unimportant

py3: fix pickling of TriangularModuleMorphism

Originally reported broken in trac #17957, it actually already worked on Python
3 which has better support for pickling built-in methods.  However, the
original problem can also be fixed easily with a little bit of refactoring.

py3: fix pickling of ModuleMorphismFromMatrix

Similarly to TriangularModuleMorphism, this already worked on Python 3, but
the original underlying bug could be fixed on Python 2 easily with some
refactoring.

I'm not entirely happy with how ModuleMorphism handles _on_basis in general
but this okay for now.

py3: add a necessary int() for the arguments to islice

on Python 3.7 we won't need this anymore

fix what we agreed is likely a typo in this test

see https://groups.google.com/d/msg/sage-devel/6YPo-Ny13y8/NetK5g5yBgAJ

py3: replace testing exact hash value with something else

these tests are still not great, though it's fine enough

misc whitespace and other delinting