Trac #26639: use PEP 289 generator expressions in geometry/cone.py.
A generator expression in Python is a more general version of a list
comprehension. The problem with list comprehensions is that, well,
they create lists. And often, we don't need the list for anything
other than to traverse it. In that case, an iterator would be more
efficient, although the list comprehension syntax is pretty convenient.

That's where generator expressions come in. The following code creates
a list of unit-norm generators for the cone K, and then sums its entries:

  sage: K = Cone([(2,0), (0,2)])
  sage: sum([ vector(QQbar,g).normalized() for g in K ])

However, we don't need that list: we just want the sum. In a generator
expression, the list brackets are replaced by parentheses, and an
iterator (as opposed to a list) is returned. In many cases, this is a
more efficient drop-in replacement for a list comprehension. For
example,

  sage: K = Cone([(2,0), (0,2)])
  sage: sum(( vector(QQbar,g).normalized() for g in K ))

This commit fixes a few "easy" cases where list comprehensions were
used, but where generator expressions should work too.