Trac #26623: add cones.<foo> shortcuts for common cones.
Add a new module sage.geometry.cone_catalog that allows users to
easily construct a few polyhedral cones:

  * The nonnegative orthant,
  * The rearrangement cone,
  * The Schur cone,
  * The trivial cone.

Inside sage.geometry.all, the module is lazily imported as "cones" so
that the functions it defines are available in the global namespace,
but are only imported if necessary. Using it is as simple as

  sage: cones.nonnegative_orthant(3)

This functionality is in the same spirit as the pre-defined "graphs"
object that lets one create the Petersen graph with graphs.PetersenGraph().