some fixes to doc parsing
in particular, when extracting strings from the GAP documentation (e.g. for
use in docstrings, as opposed to passing to the pager) extract the correct
section of the documentation for a given object, rather than the entire
document that that object is documented in (see e.g. the test for
SymmetricGroup.__doc__ which nows starts with the docs for SymmetricGroup
rather than the entire "50 Group Libraries"