Allow Gap.__getattr__ to return GAP objects other than just functions
There's no immediately obvious reason it should only be able to return
functions; there are some GAP globals that are not functions that might
still be of interest.

Cleans up the common_gap_functions list to remove entries that are either
not actually functions, or that no longer exist.  Added a common_gap_globals
list which includes common_gap_functions, but also other common globals
previously from common_gap_functions but that are not actually functions.