initial DESTDIR support for the gcc SPKG
Adds a spkg-preinst script to uninstall gfortran if it's installed, and moved
the bits that modify $SAGE_LOCAL after installation (e.g. forcing rebuilds of
packages) into spkg-postinst, as well as spkg-postrm for when uninstalling gcc.
This does not yet deal with some of the OSX-specific hacks in the spkg-install.

Also includes gcc-multilib-multiarch.patch, which ensures that gcc's libraries
are installed in lib/ instead of lib64/ on Linux.