Modify algorithm to generate random lattices.
This modification ensures that all lattices of the given size are
generated with positive probability (as long as 0 < p < 1).

See https://trac.sagemath.org/ticket/27130 for details.

To generate the example given in the ticket of a lattice that was
impossible to generate:

set_random_seed(15560)
L = posets.RandomLattice(9, 0.9)

The examples given in the documentation for RandomLattice() and
_random_lattice() produce the same results as before.