Move all the --with-mp logic into the spkg-configure.m4 for MPIR so that it is local to where it is used
Also, most of the logic is in mpir/spkg-configure.m4 rather than
gmp/spkg-configure.m4 just because MPIR is the implementation we install by
default if it can't be used from the sytem.  It could go either way, but this
way seemed a bit more logical for that reason.

Then, the gmp/spkg-configure.m4 is mostly just a stub.

Finally, uses the symbol __gmpq_cmp_z to check for GMP>=6.1.0 (or whatever the
equivalent MPIR is).