Trac #27354: replace one use of var() with SR.var() in a doctest.
While working on the doctests for the assume() function, I noticed
that one example uses the var() function to inject symbolic variables
into global scope. That is considered a bad programming practice, and
we have a better alternative: to assign the return value of SR.var()
instead. This commit updates that one doctest to use SR.var() instead
of var().