Ticket #27514: Rename the default implementation of build_many to _build_many and always define it at module-level.
This way even on Cygwin its default implementation is not overshadowed
and it can still be tested directly.  This fixes the otherwise broken
test in sage_setup.docbuild (note: that test is a regression test for an
issue that is not even applicable to the new _build_many implementation
added for Cygwin).