move some combinat references to master reference file
The "name mangling" performed by this class is a workaround to
pickling nested classes which is a non-issue on python 3, so there
is no reason to do it.

Move this reference to the module level.

This is useful in the direction of https://trac.sagemath.org/ticket/27497 but is
now also needed for the docs to build properly.

Fix up sage_autodoc to work with normal Python 3 nested classes with qualnames,
as opposed to Python 2 with Sage's mangled names for nested classes.

trac 27692: pyflakes cleanup in sage_autodoc.py

trac 27692: move some references to master bibliography file