Trac #27814: add configure check for system "rw" library.
This adds the usual spkg-configure.m4 for the "rw" package. Our graph
stuff uses a few functions from this library that are present in the
"rw.h" header. The check looks for that header (AC_CHECK_HEADER), and
makes sure that at least one function we expect to be in the library
is in fact there (AC_SEARCH_LIBS).